Carve out observability_options test
